% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{attr_or_null}
\alias{attr_or_null}
\title{Retrieve Attribute or Return NULL if Missing}
\usage{
attr_or_null(.data, attr_name)
}
\arguments{
\item{.data}{An object to inspect.}

\item{attr_name}{A string. The name of the attribute to retrieve.}
}
\value{
The value of the specified attribute, or \code{NULL} if not present.
}
\description{
A safe variant of \code{attr_or_abort()} that returns \code{NULL} instead of throwing an error
when the specified attribute does not exist.
}
\examples{
obj <- structure(list(a = 1), attr = list(my_attr = "value"))
attr_or_null(obj, "my_attr")
#> "value"

attr_or_null(obj, "missing")
#> NULL

}
\keyword{internal}
