% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_checks_missing.R
\name{calculate_completeness_summary}
\alias{calculate_completeness_summary}
\title{Summarize Data Completeness by Year}
\usage{
calculate_completeness_summary(
  .data,
  admin_level = c("national", "adminlevel_1", "district"),
  region = NULL
)
}
\arguments{
\item{.data}{A \code{cd_data} object.}

\item{admin_level}{One of \code{'national'}, \code{'adminlevel_1'}, or \code{'district'}.}
}
\value{
A \code{cd_completeness_summary} tibble:
\itemize{
\item One row per year and group
\item \verb{mis_<indicator>} columns with \% non-missing values
\item \code{mean_mis_all} summarizing all indicators
}
}
\description{
Computes yearly percentage of non-missing values for each indicator, along with
overall completeness summaries.
}
\examples{
\dontrun{
calculate_completeness_summary(data)
}

}
