% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_threshold.R
\name{calculate_threshold}
\alias{calculate_threshold}
\title{Calculate Dropout Coverage for Health Indicators Below a Threshold}
\usage{
calculate_threshold(
  .data,
  denominator = c("dhis2", "anc1", "penta1", "penta1derived"),
  indicator = c("maternal", "vaccine")
)
}
\arguments{
\item{.data}{A tibble of class \code{cd_data}.}

\item{indicator}{Character. The specific health indicator to evaluate. Options are:
\itemize{
\item \code{"coverage"}:coverage indicators.
\item \code{"dropout"}: dropout indicators.
}}

\item{admin_level}{The level of analysis.}

\item{sbr}{Numeric. The stillbirth rate. Default is \code{0.02}.}

\item{nmr}{Numeric. Neonatal mortality rate. Default is \code{0.025}.}

\item{pnmr}{Numeric. Post-neonatal mortality rate. Default is \code{0.024}.}

\item{anc1survey}{Numeric. Survey-derived coverage rate for ANC-1 (antenatal care, first visit). Default is \code{0.98}.}

\item{dpt1survey}{Numeric. Survey-derived coverage rate for Penta-1 (DPT1 vaccination). Default is \code{0.97}.}

\item{survey_year}{Integer. The year of Penta-1 survey provided}

\item{preg_loss}{Numeric. Pregnancy loss rate}

\item{twin}{Numeric. Twin birth rate. Default is \code{0.015}.}
}
\value{
A tibble with the selected administrative level and coverage value for regions
that do not meet the below-10\% threshold for the specified indicator and year. If no regions
meet the criteria, a default row is returned with "None" and 0 as values.
}
\description{
This function filters health indicator data to identify the percentage of administrative
regions where the coverage for a specified indicator falls below a 10\% threshold for a given year.
If no regions meet the criteria (i.e., all values are below the threshold), a default output is returned.
}
\examples{
\dontrun{
# Example usage:
result <- calculate_threshold(data, filter_year = 2023, indicator = "zerodose", source = "dhis2")
result
}

}
