% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_mortality.R
\name{create_mortality_ratios}
\alias{create_mortality_ratios}
\title{Extract Latest Mortality Ratios for UN Comparison}
\usage{
create_mortality_ratios(
  .data,
  mortality_data,
  admin_level = c("national", "adminlevel_1", "district")
)
}
\arguments{
\item{.data}{The main mortality dataset.}

\item{mortality_data}{A \code{cd_un_mortality} object.}
}
\value{
A tibble of class \code{cd_mortality_ratio} with latest available values and UN means.
}
\description{
Joins computed national mortality indicators with UN estimates, returns latest value per indicator.
}
