% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_mortality.R
\name{create_mortality_summary}
\alias{create_mortality_summary}
\title{Create Mortality Summary Object}
\usage{
create_mortality_summary(.data)
}
\arguments{
\item{.data}{An object of class \code{cd_data}.}
}
\value{
A \code{cd_mortality_rare} tibble.
}
\description{
Combines national and admin1 summaries and formats as a custom class for plotting.
}
