% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fpet_summary.R
\name{generate_fpet_summary}
\alias{generate_fpet_summary}
\title{Load and Generate FPET Output}
\usage{
generate_fpet_summary(path)
}
\arguments{
\item{path}{A Character for the path of the FPET input file}
}
\value{
A \code{cd_fpet} object.
}
\description{
Processes raw FPET data (e.g., from Track20 or UNFPA) into a structured tibble
for visualization and interpretation of modern contraceptive prevalence and
demand satisfied.
}
