% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_indicator_without_opd_ipd}
\alias{get_indicator_without_opd_ipd}
\title{Get Indicators excluding indicators without denominator}
\usage{
get_indicator_without_opd_ipd()
}
\value{
Character vector of all indicators
}
\description{
Flatten all indicators from all groups
}
