% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{get_plot_admin_column}
\alias{get_plot_admin_column}
\title{Determine the Administrative Column to Use for Plotting}
\usage{
get_plot_admin_column(admin_level, region = NULL)
}
\arguments{
\item{admin_level}{A string. Either \code{"adminlevel_1"} or \code{"district"}.}

\item{region}{Optional. A region name or \code{NULL}. If specified, districts will be used for plotting.}
}
\value{
A string: either \code{"district"} or \code{"adminlevel_1"}.
}
\description{
Returns the appropriate administrative unit column to use on plots based on the selected
administrative level and whether a region is specified. Used to decide whether to facet or
label plots by district or by adminlevel_1.
}
\examples{
get_plot_admin_column("adminlevel_1")
#> "adminlevel_1"

get_plot_admin_column("adminlevel_1", region = "Nairobi")
#> "district"

get_plot_admin_column("district")
#> "district"

}
