% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fpet_summary.R
\name{interpret.cd_fpet}
\alias{interpret.cd_fpet}
\title{Interpret FPET Trends}
\usage{
\method{interpret}{cd_fpet}(x, ...)
}
\arguments{
\item{x}{A \code{cd_fpet} object.}

\item{...}{Unused.}
}
\value{
A character string summarizing the trend for each indicator.
}
\description{
Provides a brief textual summary of FPET trends by indicator over time.
}
