% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cd-data.R
\name{is_maternal_indicator}
\alias{is_maternal_indicator}
\title{Detemine if an indicator is maternal indicator}
\usage{
is_maternal_indicator(indicator)
}
\value{
TRUE if and indicator is maternal
}
\description{
Default grouping of indicators used in CD2030 coverage framework
}
