% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fpet_summary.R
\name{plot.cd_fpet}
\alias{plot.cd_fpet}
\title{Plot FPET Estimates}
\usage{
\method{plot}{cd_fpet}(x, ...)
}
\arguments{
\item{x}{A \code{cd_fpet} object.}

\item{...}{Additional arguments for customization (not used).}
}
\value{
A ggplot2 object.
}
\description{
Creates a line plot for modern contraceptive prevalence rate (mCPR) over time,
optionally faceted by subnational region.
}
