% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.service_utlization.R
\name{plot.cd_service_utilization}
\alias{plot.cd_service_utilization}
\title{Plot Service Utilization Indicators}
\usage{
\method{plot}{cd_service_utilization}(x, plot_type = c("opd", "ipd", "under5", "cfr", "deaths"), region = NULL)
}
\arguments{
\item{x}{A \code{cd_service_utilization} object created by \code{\link[=compute_service_utilization]{compute_service_utilization()}}.}

\item{plot_type}{One of:
\itemize{
\item \code{"opd"}: OPD visits per person
\item \code{"ipd"}: IPD admissions per person
\item \code{"under5"}: Under-5 share of OPD and IPD
\item \code{"cfr"}: Case fatality rate (under-5 and total)
\item \code{"deaths"}: Proportion of deaths that are under-5
}}

\item{region}{Character description}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Visualizes service utilization over time from a \code{cd_service_utilization} object.
}
\examples{
\dontrun{
plot(compute_service_utilization(dat), plot_type = "opd")
}

}
